created all lines, also figured out number of pixels per memter and found distance for all lines. Next, I need to look at pathfinding

+++ app/src/main/java/com/majorproject/caverouteplanner/ui/components/SurveyPath.kt
@@ -3,7 +3,7 @@ package com.majorproject.caverouteplanner.ui.components
 data class SurveyPath(
     var id: Int,
     var ends: Pair<Int, Int>,
-    var distance: Double = 0.0,
+    var distance: Float = 0.0f,
     var hasWater: Boolean = false,
     var altitude: Int = 0,
     var isHardTraverse: Boolean = false
